Parallelize the extension
Currently the extension forces the sphinx build to be sequential,
('parallel_{read,write}_safe': False). This is because it needs
to collect the links from all docs in the build together, and
the simplest way is to process all docs in the same process.

For parallelising the read stage, Sphinx provides the 'env-merge-info'
event, which can be hooked to merge together results from various
read workers. Unfortunately link processing happens in the builder
stage, which comes afterwards and provides no similar event.

But we can work around this by storing the links in a data structure
shared between the worker processes. In this patch, we opt for
multiprocessing.Manager.